Extract spdy disconnect test; update mockito
