+++ modules/http/pom.xml
@@ -53,6 +53,11 @@
     <packaging>bundle</packaging>
     <version>2.3.36-MULE-026-SNAPSHOT</version>
     <name>grizzly-http</name>
+
+    <properties>
+        <mockitoVersion>4.11.0</mockitoVersion>
+    </properties>
+
     <build>
         <defaultGoal>install</defaultGoal>
         <resources>
@@ -105,9 +110,9 @@
         </dependency>
         <dependency>
             <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>1.9.5</version>
+            <artifactId>mockito-inline</artifactId>
             <scope>test</scope>
+            <version>${mockitoVersion}</version>
         </dependency>
     </dependencies>
 </project>
