+++ modules/spdy/src/test/java/org/glassfish/grizzly/spdy/NIOInputSourcesTest.java
@@ -503,110 +503,6 @@ public class NIOInputSourcesTest extends AbstractSpdyTest {
 
     }
 
-    /**
-     * Test ReadHandler.onError to be notified, when client unexpectedly
-     * terminates the connection
-     */
-    @SuppressWarnings({"unchecked"})
-    @Test
-    public void testDisconnect() throws Throwable {
-
-        final AtomicInteger bytesRead = new AtomicInteger();
-        final FutureImpl<Integer> resultFuture = SafeFutureImpl.create();
-
-        final TCPNIOTransport clientTransport = TCPNIOTransportBuilder.newInstance().build();
-        clientTransport.setProcessor(
-                createClientFilterChain(spdyVersion, spdyMode, isSecure));
-        
-        final HttpHandler httpHandler = new HttpHandler() {
-
-            @Override
-            public void service(final Request request,
-                    final Response response) throws Exception {
-                response.suspend();
-                final NIOInputStream inputStream = (NIOInputStream) request.getInputStream();
-
-                inputStream.notifyAvailable(new ReadHandler() {
-
-                    @Override
-                    public void onDataAvailable() throws IOException {
-                        final int readyData = inputStream.readyData();
-                        inputStream.skip(readyData);
-                        bytesRead.addAndGet(readyData);
-
-                        inputStream.notifyAvailable(this);
-                    }
-
-
-                    @Override
-                    public void onAllDataRead() throws IOException {
-                        final int readyData = inputStream.readyData();
-                        inputStream.skip(readyData);
-                        bytesRead.addAndGet(readyData);
-                        resultFuture.failure(new IllegalStateException("Connection should have been terminated"));
-
-                        response.resume();
-                    }
-
-                    @Override
-                    public void onError(Throwable t) {
-                        resultFuture.failure(t);
-                        
-                        response.resume();
-                    }
-                });
-            }
-
-        };
-
-        final HttpServer server = createWebServer(httpHandler);
-
-        try {
-            server.start();
-            clientTransport.start();
-
-            Future<Connection> connectFuture = clientTransport.connect("localhost", PORT);
-            Connection connection = null;
-            try {
-                connection = connectFuture.get(10, TimeUnit.SECONDS);
-                HttpRequestPacket packet = (HttpRequestPacket) createRequest(PORT, "POST", null, null);
-                packet.setContentLength(5000);
-                connection.write(packet);
-
-                HttpContent content = HttpContent.builder(packet).content(
-                        Buffers.wrap(null, buildString(2500))).build();
-
-                connection.write(content, new EmptyCompletionHandler<WriteResult>() {
-
-                    @Override
-                    public void completed(WriteResult result) {
-                        result.getConnection().closeSilently();
-                    }
-                });
-
-                try {
-                    final Integer i = resultFuture.get(10, TimeUnit.SECONDS);
-                    fail("Wrapped EOFException expected");
-                } catch (ExecutionException e) {
-                    assertEquals("NOT EOF Exception:\n" +
-                            Exceptions.getStackTraceAsString(e.getCause()),
-                            EOFException.class, e.getCause().getClass());
-                }
-            } finally {
-                // Close the client connection
-                if (connection != null) {
-                    connection.closeSilently();
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-            fail();
-        } finally {
-            clientTransport.shutdownNow();
-            server.shutdownNow();
-        }
-    }
-
     // --------------------------------------------------------- Private Methods
